+++ src/Generator/Config/FederationRepresentationResolverGeneratorConfigInterface.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Axtiva\FlexibleGraphql\Generator\Config;
+
+interface FederationRepresentationResolverGeneratorConfigInterface extends ModelGeneratorConfigInterface
+{}